Refactor: remove vestigial UkuleleString typealias from FretboardViewModel
The typealias was added when UkuleleString moved to the shared domain
package as a backward-compatibility shim. All 15 callers now import
directly from com.baijum.ukufretboard.domain.UkuleleString.

Fixes #403
