+++ app/src/main/java/com/baijum/ukufretboard/ui/FavoritesTab.kt
@@ -57,7 +57,7 @@ import com.baijum.ukufretboard.data.FavoriteVoicing
 import com.baijum.ukufretboard.data.Notes
 import com.baijum.ukufretboard.domain.ChordVoicing
 import com.baijum.ukufretboard.viewmodel.FavoritesViewModel
-import com.baijum.ukufretboard.viewmodel.UkuleleString
+import com.baijum.ukufretboard.domain.UkuleleString
 import sh.calvin.reorderable.ReorderableItem
 import sh.calvin.reorderable.rememberReorderableLazyGridState
