+++ app/src/main/java/com/baijum/ukufretboard/viewmodel/ChordLibraryViewModel.kt
@@ -11,6 +11,7 @@ import com.baijum.ukufretboard.domain.ChordNameParser
 import com.baijum.ukufretboard.domain.ChordSearchResult
 import com.baijum.ukufretboard.domain.ChordVoicing
 import com.baijum.ukufretboard.domain.Transpose
+import com.baijum.ukufretboard.domain.UkuleleString
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
